requirements.txt
# Core dependencies
fastapi[all]>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# AI and ML dependencies
openai>=1.3.0
numpy>=1.24.0
pipecat-ai[daily,cartesia,openai,silero]>=0.1.0
pipecat-ai-flows>=0.1.0
mlx-whisper>=0.1.0

# Logging
loguru>=0.7.0

# Async HTTP
aiohttp>=3.9.0

# Type checking
typing-extensions>=4.8.0

# WebSocket support
websockets>=12.0

# Optional development dependencies
pytest>=7.4.0  # for testing
black>=23.11.0  # for code formatting
isort>=5.12.0  # for import sorting

